+++ railties/test/commands/credentials_test.rb
@@ -170,6 +170,36 @@ class Rails::Command::CredentialsTest < ActiveSupport::TestCase
     assert_match %r/foo: bar: bad/, run_edit_command
   end
 
+  test "edit command works with git conflicts if internalise_conflictsinternalise_conflicts is enabled" do
+    run_edit_command(environment: "development")
+
+    write_credentials "baz: foo\nfoo: bar\nbar: baz", environment: "development"
+    left_file = File.read(app_path("config", "credentials", "development.yml.enc"))
+
+    write_credentials "baz: foo\nfoo: baz\nbar: baz", environment: "development"
+    right_file = File.read(app_path("config", "credentials", "development.yml.enc"))
+
+    merge_conflict = <<~CONFLICT
+      <<<<<<< HEAD
+      #{left_file}
+      =======
+      #{right_file}
+      >>>>>>> @{-1}
+    CONFLICT
+    File.write(app_path("config", "credentials", "development.yml.enc"), merge_conflict)
+
+    decrypted_conflict = <<~CONFLICT
+      baz: foo
+      <<<<<<< HEAD
+      foo: bar
+      =======
+      foo: baz
+      >>>>>>> @{-1}
+      bar: baz
+    CONFLICT
+
+    assert_includes(run_edit_command(environment: "development", internalise_conflicts: true), decrypted_conflict.strip)
+  end
 
   test "show credentials" do
     assert_match DEFAULT_CREDENTIALS_PATTERN, run_show_command
@@ -353,10 +383,12 @@ class Rails::Command::CredentialsTest < ActiveSupport::TestCase
   private
     DEFAULT_CREDENTIALS_PATTERN = /access_key_id: 123\n.*secret_key_base: \h{128}\n/m
 
-    def run_edit_command(visual: "cat", editor: "cat", environment: nil, **options)
+    def run_edit_command(visual: "cat", editor: "cat", environment: nil, internalise_conflicts: nil, **options)
       switch_env("VISUAL", visual) do
         switch_env("EDITOR", editor) do
-          args = environment ? ["--environment", environment] : []
+          args = []
+          args << ["--environment", environment] if environment
+          args << ["--internalise-conflicts"] if internalise_conflicts
           rails "credentials:edit", args, **options
         end
       end